refactor: scheduler class 생성하여 schedule 호ìe 호출 기능 분리 #88
